Rename components to pascal-case.
