+++ src/components/common/Logo.tsx
@@ -11,12 +11,12 @@ export function Logo ({className, ...props}: React.ComponentProps<"svg">) {
             <path fill="var(--foreground)" d="M28.608 42.152c2.39.427 4.416 1.685 6.08 3.776 1.664 2.09 2.496 4.459 2.496 7.104 0 2.261-.597 4.31-1.792 6.144-1.152 1.792-2.837 3.221-5.056 4.288-2.219 1.024-4.8 1.536-7.744 1.536H4.8V20.52h16.96c3.03 0 5.632.512 7.808 1.536s3.819 2.41 4.928 4.16c1.11 1.707 1.664 3.627 1.664 5.76 0 2.56-.683 4.693-2.048 6.4-1.365 1.707-3.2 2.965-5.504 3.776Zm-16.512-2.944h9.024c2.39 0 4.245-.533 5.568-1.6 1.365-1.11 2.048-2.688 2.048-4.736 0-2.005-.683-3.563-2.048-4.672-1.323-1.152-3.179-1.728-5.568-1.728h-9.024v12.736Zm9.856 19.84c2.475 0 4.416-.597 5.824-1.792 1.408-1.195 2.112-2.859 2.112-4.992 0-2.176-.747-3.904-2.24-5.184-1.493-1.28-3.477-1.92-5.952-1.92h-9.6v13.888h9.856Zm20.729-11.84c0-3.541.725-6.677 2.175-9.408 1.494-2.73 3.5-4.843 6.017-6.336 2.56-1.536 5.376-2.304 8.448-2.304 2.773 0 5.183.555 7.231 1.664 2.091 1.067 3.755 2.41 4.992 4.032v-5.12h7.36V65h-7.36v-5.248c-1.237 1.664-2.922 3.05-5.055 4.16-2.134 1.11-4.566 1.664-7.297 1.664-3.029 0-5.802-.768-8.32-2.304-2.517-1.579-4.522-3.755-6.016-6.528-1.45-2.816-2.175-5.995-2.175-9.536Zm28.864.128c0-2.432-.513-4.544-1.537-6.336-.98-1.792-2.282-3.157-3.903-4.096a10.297 10.297 0 0 0-5.249-1.408c-1.877 0-3.626.47-5.248 1.408-1.62.896-2.944 2.24-3.967 4.032-.982 1.75-1.472 3.84-1.472 6.272 0 2.432.49 4.565 1.472 6.4 1.023 1.835 2.346 3.243 3.968 4.224 1.663.939 3.413 1.408 5.247 1.408 1.878 0 3.627-.47 5.248-1.408 1.622-.939 2.923-2.304 3.904-4.096 1.025-1.835 1.537-3.968 1.537-6.4ZM106.28 29.16c2.773 0 5.248.576 7.424 1.728 2.218 1.152 3.946 2.859 5.184 5.12 1.237 2.261 1.856 4.992 1.856 8.192V65h-7.232V45.288c0-3.157-.79-5.568-2.368-7.232-1.579-1.707-3.734-2.56-6.464-2.56-2.731 0-4.907.853-6.529 2.56-1.578 1.664-2.367 4.075-2.367 7.232V65h-7.297V29.736h7.297v4.032c1.194-1.45 2.709-2.581 4.544-3.392a14.852 14.852 0 0 1 5.952-1.216Zm37.722 18.24L160.258 65h-9.856l-13.056-15.168V65h-7.296V17.64h7.296v27.52l12.8-15.424h10.112L144.002 47.4Zm50.222-26.88v5.952h-11.84V65h-7.296V26.472h-11.904V20.52h31.04Zm22.685 45.056c-3.328 0-6.336-.747-9.024-2.24-2.688-1.536-4.8-3.67-6.336-6.4-1.536-2.773-2.304-5.973-2.304-9.6 0-3.584.789-6.763 2.368-9.536 1.578-2.773 3.733-4.907 6.464-6.4 2.73-1.493 5.781-2.24 9.152-2.24 3.37 0 6.421.747 9.152 2.24 2.73 1.493 4.885 3.627 6.464 6.4 1.578 2.773 2.368 5.952 2.368 9.536 0 3.584-.811 6.763-2.432 9.536-1.622 2.773-3.84 4.928-6.656 6.464-2.774 1.493-5.846 2.24-9.216 2.24Zm0-6.336a10.68 10.68 0 0 0 5.248-1.344c1.664-.896 3.008-2.24 4.032-4.032 1.024-1.792 1.536-3.968 1.536-6.528 0-2.56-.491-4.715-1.472-6.464-.982-1.792-2.283-3.136-3.904-4.032a10.68 10.68 0 0 0-5.248-1.344c-1.878 0-3.627.448-5.248 1.344-1.579.896-2.838 2.24-3.776 4.032-.939 1.75-1.408 3.904-1.408 6.464 0 3.797.96 6.741 2.88 8.832 1.962 2.048 4.416 3.072 7.36 3.072Zm40.812 6.336c-3.328 0-6.336-.747-9.024-2.24-2.688-1.536-4.8-3.67-6.336-6.4-1.536-2.773-2.304-5.973-2.304-9.6 0-3.584.789-6.763 2.368-9.536 1.579-2.773 3.733-4.907 6.464-6.4 2.731-1.493 5.781-2.24 9.152-2.24 3.371 0 6.421.747 9.152 2.24 2.731 1.493 4.885 3.627 6.464 6.4 1.579 2.773 2.368 5.952 2.368 9.536 0 3.584-.811 6.763-2.432 9.536a17.004 17.004 0 0 1-6.656 6.464c-2.773 1.493-5.845 2.24-9.216 2.24Zm0-6.336c1.877 0 3.627-.448 5.248-1.344 1.664-.896 3.008-2.24 4.032-4.032 1.024-1.792 1.536-3.968 1.536-6.528 0-2.56-.491-4.715-1.472-6.464-.981-1.792-2.283-3.136-3.904-4.032a10.677 10.677 0 0 0-5.248-1.344c-1.877 0-3.627.448-5.248 1.344-1.579.896-2.837 2.24-3.776 4.032-.939 1.75-1.408 3.904-1.408 6.464 0 3.797.96 6.741 2.88 8.832 1.963 2.048 4.416 3.072 7.36 3.072Z"/>
             <defs>
                 <linearGradient id="a" x1="270" x2="250.883" y1="56.5" y2="53.795" gradientUnits="userSpaceOnUse">
-                    <stop stop-color="var(--primary)"/>
-                    <stop offset="1" stop-color="var(--secondary)"/>
+                    <stop stopColor="var(--primary)"/>
+                    <stop offset="1" stopColor="var(--secondary)"/>
                 </linearGradient>
                 <linearGradient id="b" x1="226.5" x2="209.678" y1="56" y2="53.899" gradientUnits="userSpaceOnUse">
-                    <stop stop-color="var(--primary)"/>
-                    <stop offset="1" stop-color="var(--secondary)"/>
+                    <stop stopColor="var(--primary)"/>
+                    <stop offset="1" stopColor="var(--secondary)"/>
                 </linearGradient>
             </defs>
         </svg>
