+++ src/components/ui/skeleton.tsx
@@ -5,7 +5,7 @@ function Skeleton({ className, ...props }: React.ComponentProps<"div">) {
   return (
     <div
       data-slot="skeleton"
-      className={cn("bg-primary/10 animate-pulse rounded-md", className)}
+      className={cn("bg-muted/50 animate-pulse rounded-md", className)}
       {...props}
     />
   )
