+++ src/index.css
@@ -181,9 +181,6 @@
 
     --color-sidebar-accent: var(--sidebar-accent);
 
-    --color-sidebar-primary-foreground: var(--sidebar-primary-foreground);
-
-    --color-sidebar-primary: var(--sidebar-primary);
 
     --color-sidebar-foreground: var(--sidebar-foreground);
